embeddings/requirements.txt
faiss-cpu==1.12.0
websockets==15.0.1
